Add getOrFirst method to ParameterBag, add precision to timeElapsed method

+++ src/SystemInfo.php
@@ -14,7 +14,7 @@ class SystemInfo
         return count(get_declared_classes());
     }
 
-    public function timeElapsed(): int
+    public function timeElapsed(): float
     {
         return microtime(true) - FRAMEWORK_START;
     }
